Moved engine connection from webapp to cli
