+++ backend-diagram-converter/cli/pom.xml
@@ -29,6 +29,14 @@
       <artifactId>commons-io</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.springframework</groupId>
+      <artifactId>spring-web</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.fasterxml.jackson.core</groupId>
+      <artifactId>jackson-databind</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.junit.jupiter</groupId>
       <artifactId>junit-jupiter</artifactId>
       <scope>test</scope>
@@ -73,7 +81,7 @@
               <archive>
                 <manifest>
                   <addClasspath>true</addClasspath>
-                  <mainClass>org.camunda.community.converter.cli.Main</mainClass>
+                  <mainClass>org.camunda.community.converter.cli.ConvertCommand</mainClass>
                 </manifest>
               </archive>
             </configuration>
