+++ backend-diagram-converter/cli/src/main/java/org/camunda/community/converter/cli/ProcessDefinitionDto.java
@@ -1,4 +1,4 @@
-package org.camunda.community.converter.webapp.dto;
+package org.camunda.community.converter.cli;
 
 public class ProcessDefinitionDto {
   private String id;
