+++ backend-diagram-converter/webapp/src/main/java/org/camunda/community/converter/webapp/BpmnConverterService.java
@@ -1,13 +1,7 @@
 package org.camunda.community.converter.webapp;
 
-import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.StringWriter;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.stream.Collectors;
-import org.camunda.bpm.model.bpmn.Bpmn;
 import org.camunda.bpm.model.bpmn.BpmnModelInstance;
 import org.camunda.bpm.model.xml.instance.DomDocument;
 import org.camunda.community.converter.BpmnConverter;
@@ -20,13 +14,10 @@ import org.springframework.stereotype.Service;
 @Service
 public class BpmnConverterService {
   private final BpmnConverter bpmnConverter;
-  private final ProcessEngineService processEngineService;
 
   @Autowired
-  public BpmnConverterService(
-      BpmnConverter bpmnConverter, ProcessEngineService processEngineService) {
+  public BpmnConverterService(BpmnConverter bpmnConverter) {
     this.bpmnConverter = bpmnConverter;
-    this.processEngineService = processEngineService;
   }
 
   public void convert(
@@ -51,34 +42,6 @@ public class BpmnConverterService {
         modelInstance,
         appendDocumentation,
         ConverterPropertiesFactory.getInstance().merge(adaptedProperties));
-  }
-
-  public Map<String, BpmnModelInstance> convertFromEngine(boolean appendDocumentation) {
-    return processEngineService.getAllLatestBpmnXml().entrySet().stream()
-        .collect(
-            Collectors.toMap(
-                Entry::getKey,
-                e -> {
-                  BpmnModelInstance modelInstance =
-                      Bpmn.readModelFromStream(streamForString(e.getValue()));
-                  convert(modelInstance, appendDocumentation);
-                  return modelInstance;
-                }));
-  }
-
-  public List<BpmnDiagramCheckResult> checkFromEngine(boolean appendDocumentation) {
-    return processEngineService.getAllLatestBpmnXml().entrySet().stream()
-        .map(
-            e ->
-                check(
-                    e.getKey(),
-                    Bpmn.readModelFromStream(streamForString(e.getValue())),
-                    appendDocumentation))
-        .collect(Collectors.toList());
-  }
-
-  private ByteArrayInputStream streamForString(String bpmnXml) {
-    return new ByteArrayInputStream(bpmnXml.getBytes());
   }
 
   public String printXml(DomDocument document, boolean prettyPrint) {
